requirements.txt
langgraph
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
python-dotenv
streamlit
fastapi
uvicorn
langgraph-checkpoint-sqlite
tensorflow
langchain_core
python-multipart
SpeechRecognition
pyaudio
streamlit-webrtc
pydub
pipwin
chromadb
pypdf2
python-docx
openpyxl
googletrans
sentence-transformers
langchain-huggingface
langchain-openai
langchain-pinecone
pinecone-client>=3.0.0
langchain-community
langchain-text-splitters
pypdf
docx2txt